Move stage generator into individual namespace.
