+++ osu.Game.Rulesets.Karaoke.Tests/Editor/Generator/Stages/BaseStageElementCategoryGeneratorTest.cs
@@ -5,11 +5,11 @@ using System;
 using NUnit.Framework;
 using osu.Game.Rulesets.Karaoke.Beatmaps;
 using osu.Game.Rulesets.Karaoke.Edit.Generator;
-using osu.Game.Rulesets.Karaoke.Edit.Generator.Beatmaps.Stages;
+using osu.Game.Rulesets.Karaoke.Edit.Generator.Stages;
 using osu.Game.Rulesets.Karaoke.Objects;
 using osu.Game.Rulesets.Karaoke.Stages;
 
-namespace osu.Game.Rulesets.Karaoke.Tests.Editor.Generator.Beatmaps.Stages;
+namespace osu.Game.Rulesets.Karaoke.Tests.Editor.Generator.Stages;
 
 public abstract class BaseLyricStageElementCategoryGeneratorTest<TGenerator, TObject, TStageElement, TConfig>
     : BaseStageElementCategoryGeneratorTest<TGenerator, TObject, TStageElement, Lyric, TConfig>
