+++ osu.Game.Rulesets.Karaoke.Tests/Editor/Generator/Stages/Classic/ClassicLyricTimingInfoGeneratorTest.cs
@@ -5,12 +5,12 @@ using System.Collections.Generic;
 using System.Linq;
 using NUnit.Framework;
 using osu.Game.Rulesets.Karaoke.Beatmaps;
-using osu.Game.Rulesets.Karaoke.Edit.Generator.Beatmaps.Stages.Classic;
+using osu.Game.Rulesets.Karaoke.Edit.Generator.Stages.Classic;
 using osu.Game.Rulesets.Karaoke.Objects;
 using osu.Game.Rulesets.Karaoke.Stages.Classic;
 using osu.Game.Rulesets.Karaoke.Tests.Helper;
 
-namespace osu.Game.Rulesets.Karaoke.Tests.Editor.Generator.Beatmaps.Stages.Classic;
+namespace osu.Game.Rulesets.Karaoke.Tests.Editor.Generator.Stages.Classic;
 
 public class ClassicLyricTimingInfoGeneratorTest
     : BaseStageInfoPropertyGeneratorTest<ClassicLyricTimingInfoGenerator, ClassicLyricTimingInfo, ClassicLyricTimingInfoGeneratorConfig>
