+++ osu.Game.Rulesets.Karaoke.Tests/Editor/Generator/Stages/Classic/ClassicStageInfoGeneratorTest.cs
@@ -4,12 +4,12 @@
 using System.Collections.Generic;
 using NUnit.Framework;
 using osu.Game.Rulesets.Karaoke.Beatmaps;
-using osu.Game.Rulesets.Karaoke.Edit.Generator.Beatmaps.Stages.Classic;
+using osu.Game.Rulesets.Karaoke.Edit.Generator.Stages.Classic;
 using osu.Game.Rulesets.Karaoke.Objects;
 using osu.Game.Rulesets.Karaoke.Stages.Classic;
 using osu.Game.Rulesets.Karaoke.Tests.Helper;
 
-namespace osu.Game.Rulesets.Karaoke.Tests.Editor.Generator.Beatmaps.Stages.Classic;
+namespace osu.Game.Rulesets.Karaoke.Tests.Editor.Generator.Stages.Classic;
 
 public class ClassicStageInfoGeneratorTest : BaseStageInfoGeneratorTest<ClassicStageInfoGenerator, ClassicStageInfo, ClassicStageInfoGeneratorConfig>
 {
