+++ osu.Game.Rulesets.Karaoke/Edit/Generator/Stages/Classic/ClassicLyricLayoutCategoryGenerator.cs
@@ -9,7 +9,7 @@ using osu.Game.Rulesets.Karaoke.Beatmaps;
 using osu.Game.Rulesets.Karaoke.Objects;
 using osu.Game.Rulesets.Karaoke.Stages.Classic;
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Generator.Beatmaps.Stages.Classic;
+namespace osu.Game.Rulesets.Karaoke.Edit.Generator.Stages.Classic;
 
 public class ClassicLyricLayoutCategoryGenerator : StageInfoPropertyGenerator<ClassicLyricLayoutCategory, ClassicLyricLayoutCategoryGeneratorConfig>
 {
