+++ osu.Game.Rulesets.Karaoke/Edit/Generator/Stages/Classic/ClassicLyricTimingInfoGeneratorConfig.cs
@@ -3,9 +3,9 @@
 
 using osu.Framework.Bindables;
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Generator.Beatmaps.Stages.Classic;
+namespace osu.Game.Rulesets.Karaoke.Edit.Generator.Stages.Classic;
 
-public class ClassicStageInfoGeneratorConfig : StageInfoGeneratorConfig
+public class ClassicLyricTimingInfoGeneratorConfig : GeneratorConfig
 {
     /// <summary>
     /// How may lyric can be in the stage at the same time.
