+++ osu.Game.Rulesets.Karaoke/Edit/Generator/Stages/Preview/PreviewStageInfoGeneratorConfig.cs
@@ -1,8 +1,8 @@
 // Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Generator.Beatmaps.Stages;
+namespace osu.Game.Rulesets.Karaoke.Edit.Generator.Stages.Preview;
 
-public class StageInfoGeneratorConfig : GeneratorConfig
+public class PreviewStageInfoGeneratorConfig : StageInfoGeneratorConfig
 {
 }
