+++ osu.Game.Rulesets.Karaoke/Edit/Generator/Stages/StageInfoGeneratorConfig.cs
@@ -1,8 +1,8 @@
 // Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Generator.Beatmaps.Stages.Preview;
+namespace osu.Game.Rulesets.Karaoke.Edit.Generator.Stages;
 
-public class PreviewStageInfoGeneratorConfig : StageInfoGeneratorConfig
+public class StageInfoGeneratorConfig : GeneratorConfig
 {
 }
