+++ osu.Game.Rulesets.Karaoke/Mods/KaraokeModClassicStage.cs
@@ -3,7 +3,7 @@
 
 using osu.Framework.Localisation;
 using osu.Game.Rulesets.Karaoke.Beatmaps;
-using osu.Game.Rulesets.Karaoke.Edit.Generator.Beatmaps.Stages.Classic;
+using osu.Game.Rulesets.Karaoke.Edit.Generator.Stages.Classic;
 using osu.Game.Rulesets.Karaoke.Stages.Classic;
 
 namespace osu.Game.Rulesets.Karaoke.Mods;
